fixed upload flow, updated reaction when playing from activity page, download reaction status file and terms&conditions api name mismatch

+++ app/src/main/java/com/banglalink/toffee/data/database/entities/ReactionStatusItem.kt
@@ -1,6 +1,8 @@
 package com.banglalink.toffee.data.database.entities
 
-import androidx.room.*
+import androidx.room.ColumnInfo
+import androidx.room.Entity
+import androidx.room.PrimaryKey
 
 @Entity(
     tableName = "reaction_status_item"
@@ -8,7 +10,7 @@ import androidx.room.*
 data class ReactionStatusItem(
     @PrimaryKey
     @ColumnInfo(name = "channel_id")
-    val channelId: Long = 0L,
+    val channelId: Int = 0,
 
     @ColumnInfo(name = "reaction_type")
     val reactionType: Int = 0,
